Summary 2
I had to do it again since I messed up. I had to work backwards since the instructions were apparently for an older version of the app, but in the cloned README, I added:
Jackson Currin, email addresse used was feckbatman4@gmail.com
I used the address feckbatman4@gmail.com and the username king-of-the-clouds
The repository I created was used my username and followed the template in the instructions.
The layout of the desktop app I'm using seems very different than what the instructions are describing, so I tried my best, and I was able to click on the repository and click "clone."